chain id now 256 cuz sepolia too high

+++ src/BaseChainSetup.sol
@@ -37,7 +37,7 @@ contract BaseChainSetup is CommonBase {
         return payable(weth);
     }
 
-    mapping(string => uint16) chainIdLookup;
+    mapping(string => uint256) chainIdLookup;
 
     function isMainnet() public returns (bool) {
         return vm.envOr("MAINNET", false) && strCompare(runtime, ENV_MAINNET);
@@ -86,7 +86,7 @@ contract BaseChainSetup is CommonBase {
     function configureChain(
         string memory chain,
         bool isGasEth,
-        uint16 chainId,
+        uint256 chainId,
         address weth
     ) public {
         try vm.createFork(_forkAlias(chain)) returns (uint256 forkId) {
